# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ShizoMusic API
  version: 0.4.0
  description: Private invitation-only ShizoMusic installation.
  license: { name: Private }
servers:
  - url: https://shizomusic.example.ts.net
security: []
paths:
...
  /auth/invitations/redeem:
    post:
      summary: Validate a one-time invitation
      operationId: redeemInvitation
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/RedeemInvitationRequest' } } } }
      responses:
        '200': { description: Invitation accepted for onboarding, content: { application/json: { schema: { $ref: '#/components/schemas/OnboardingTicket' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '410': { $ref: '#/components/responses/Error' }
  /auth/onboarding/complete:
    post:
      summary: Create a user and authorize an iPhone
      operationId: completeOnboarding
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CompleteOnboardingRequest' } } } }
      responses:
        '200': { description: Device authorized, content: { application/json: { schema: { $ref: '#/components/schemas/AuthSession' } } } }
        '401': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '410': { $ref: '#/components/responses/Error' }
  /auth/refresh:
    post:
      summary: Rotate a refresh token
      operationId: refreshSession
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/RefreshRequest' } } } }
      responses:
        '200': { description: Rotated token pair, content: { application/json: { schema: { $ref: '#/components/schemas/AuthSession' } } } }
        '401': { $ref: '#/components/responses/Error' }
  /auth/logout:
    post:
      summary: Revoke a refresh token
      operationId: logout
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/RefreshRequest' } } } }
      responses:
        '204': { description: Refresh token revoked }
  /me:
    get:
      summary: Get the current user
      operationId: currentUser
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Current user, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '401': { $ref: '#/components/responses/Error' }
  /admin/invitations:
    post:
      summary: Create a one-time member invitation
      operationId: createInvitation
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateInvitationRequest' } } } }
      responses:
        '200': { description: One-time invitation, content: { application/json: { schema: { $ref: '#/components/schemas/Invitation' } } } }
        '403': { $ref: '#/components/responses/Error' }
  /admin/devices:
    get:
      summary: List authorized devices
      operationId: listDevices
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: All authorized devices
          content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Device' } } } }
        '403': { $ref: '#/components/responses/Error' }
  /admin/devices/{deviceId}/revoke:
    post:
      summary: Revoke one device
      operationId: revokeDevice
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: deviceId, required: true, schema: { type: string, format: uuid } }]
      responses:
        '204': { description: Device and its refresh tokens revoked }
        '403': { $ref: '#/components/responses/Error' }
  /admin/users/{userId}/revoke:
    post:
      summary: Revoke a user and every device
      operationId: revokeUser
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: userId, required: true, schema: { type: string, format: uuid } }]
      responses:
        '204': { description: User and all devices revoked }
        '403': { $ref: '#/components/responses/Error' }
  /sync/push:
    post:
      summary: Push an idempotent batch of offline operations
      operationId: pushSyncOperations
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SyncPushRequest' } } }
      responses:
        '200': { description: Accepted operations and conflicts, content: { application/json: { schema: { $ref: '#/components/schemas/SyncPushResponse' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
  /sync/pull:
    get:
      summary: Pull canonical changes after an Int64 cursor
      operationId: pullSyncChanges
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: query, name: cursor, required: true, schema: { type: string, pattern: '^[0-9]+$' } }
        - { in: query, name: limit, required: false, schema: { type: integer, minimum: 1, maximum: 500, default: 200 } }
      responses:
        '200': { description: Ordered change-log page, content: { application/json: { schema: { $ref: '#/components/schemas/SyncPullResponse' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
...
  /search:
    get:
      summary: Search the shared catalog and configured music providers in parallel
      operationId: searchMusic
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: query, name: q, required: true, schema: { type: string, minLength: 1, maxLength: 200 } }
        - { in: query, name: limit, required: false, schema: { type: integer, minimum: 1, maximum: 50, default: 30 } }
        - { in: query, name: provider, required: false, schema: { type: string } }
      responses:
        '200': { description: Persisted normalized results and isolated provider failures, content: { application/json: { schema: { $ref: '#/components/schemas/MusicSearchResponse' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
  /search/providers/{provider}/retry:
    post:
      summary: Retry a single failed music provider
      operationId: retryMusicProvider
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: provider, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query: { type: string, minLength: 1, maxLength: 200 }
                limit: { type: integer, minimum: 1, maximum: 50, default: 30 }
      responses:
        '200': { description: Result from the selected provider, content: { application/json: { schema: { $ref: '#/components/schemas/MusicSearchResponse' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
  /external/audio/{provider}/{entityType}/{externalId}:
    get:
      summary: Resolve fresh provider audio and redirect the player
      operationId: resolveExternalAudio
      parameters:
        - { in: path, name: provider, required: true, schema: { type: string } }
        - { in: path, name: entityType, required: true, schema: { $ref: '#/components/schemas/ExternalEntityType' } }
        - { in: path, name: externalId, required: true, schema: { type: string } }
      responses:
        '302': { description: Fresh provider stream URL, headers: { Location: { schema: { type: string, format: uri } } } }
        '404': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '451': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /acquisitions:
    post:
      summary: Queue a server-side external media acquisition
      operationId: createAcquisition
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AcquisitionRequest' } } } }
      responses:
        '200': { description: Acquisition job, content: { application/json: { schema: { $ref: '#/components/schemas/AcquisitionJob' } } } }
        '400': { $ref: '#/components/responses/Error' }
    get:
      summary: List acquisition jobs for the current user
      operationId: listAcquisitions
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Jobs, content: { application/json: { schema: { type: object, required: [jobs], properties: { jobs: { type: array, items: { $ref: '#/components/schemas/AcquisitionJob' } } } } } } }
  /acquisitions/{id}:
    get:
      summary: Get one acquisition job
      operationId: getAcquisition
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200': { description: Job, content: { application/json: { schema: { $ref: '#/components/schemas/AcquisitionJob' } } } }
        '404': { $ref: '#/components/responses/Error' }
  /acquisitions/{id}/retry:
    post:
      summary: Retry a failed acquisition job
      operationId: retryAcquisition
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200': { description: Requeued job, content: { application/json: { schema: { $ref: '#/components/schemas/AcquisitionJob' } } } }
        '409': { $ref: '#/components/responses/Error' }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  responses:
    Error:
      description: API error
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
...
    RedeemInvitationRequest:
      type: object
      required: [code, deviceIdentifier, deviceName]
      properties:
        code: { type: string }
        deviceIdentifier: { type: string }
        deviceName: { type: string }
    OnboardingTicket:
      type: object
      required: [onboardingToken, expiresInSeconds]
      properties:
        onboardingToken: { type: string }
        expiresInSeconds: { type: integer }
    CompleteOnboardingRequest:
      type: object
      required: [onboardingToken, displayName]
      properties:
        onboardingToken: { type: string }
        displayName: { type: string, maxLength: 80 }
        avatarData: { type: [string, 'null'], contentEncoding: base64 }
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties: { refreshToken: { type: string } }
    CreateInvitationRequest:
      type: object
      required: [expiresInHours]
      properties: { expiresInHours: { type: integer, minimum: 1, maximum: 720 } }
    Invitation:
      type: object
      required: [code, expiresAt]
      properties:
        code: { type: string }
        expiresAt: { type: string, format: date-time }
    User:
      type: object
      required: [id, displayName, role]
      properties:
        id: { type: string, format: uuid }
        displayName: { type: string }
        avatarData: { type: [string, 'null'], contentEncoding: base64 }
        role: { type: string, enum: [owner, member] }
    Device:
      type: object
      required: [id, userId, name, createdAt, lastSeenAt]
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        name: { type: string }
        createdAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time }
        revokedAt: { type: [string, 'null'], format: date-time }
    AuthSession:
      allOf:
        - type: object
          required: [accessToken, refreshToken, accessExpiresAt, refreshExpiresAt, user, device]
          properties:
            accessToken: { type: string }
            refreshToken: { type: string }
            accessExpiresAt: { type: string, format: date-time }
            refreshExpiresAt: { type: string, format: date-time }
            user: { $ref: '#/components/schemas/User' }
            device: { $ref: '#/components/schemas/Device' }
    SyncEntityType: { type: string, enum: [playlist, playlistItem] }
    SyncOperationType:
      type: string
      enum: [playlist.upsert, playlist.delete, playlistItem.insert, playlistItem.move, playlistItem.remove]
    SyncOperation:
      type: object
      required: [id, entityType, entityId, operationType, payload, clientTimestamp]
      properties:
        id: { type: string, format: uuid }
        entityType: { $ref: '#/components/schemas/SyncEntityType' }
        entityId: { type: string }
        operationType: { $ref: '#/components/schemas/SyncOperationType' }
        payload: { type: string, description: JSON-encoded domain payload }
        clientTimestamp: { type: string, format: date-time }
    SyncChange:
      allOf:
        - { $ref: '#/components/schemas/SyncOperation' }
        - type: object
          required: [cursor, actorDeviceId]
          properties:
            cursor: { type: string, pattern: '^[0-9]+$' }
            actorDeviceId: { type: string, format: uuid }
    SyncConflict:
      type: object
      required: [id, operationId, entityType, entityId, reason, localPayload, serverPayload, createdAt]
      properties:
        id: { type: string, format: uuid }
        operationId: { type: string, format: uuid }
        entityType: { $ref: '#/components/schemas/SyncEntityType' }
        entityId: { type: string }
        reason: { type: string }
        localPayload: { type: string }
        serverPayload: { type: string }
        createdAt: { type: string, format: date-time }
    SyncPushRequest:
      type: object
      required: [operations]
      properties:
        operations: { type: array, maxItems: 100, items: { $ref: '#/components/schemas/SyncOperation' } }
    SyncPushResponse:
      type: object
      required: [acceptedOperationIds, cursor, conflicts]
      properties:
        acceptedOperationIds: { type: array, items: { type: string, format: uuid } }
        cursor: { type: string, pattern: '^[0-9]+$' }
        conflicts: { type: array, items: { $ref: '#/components/schemas/SyncConflict' } }
    SyncPullResponse:
      type: object
      required: [changes, cursor, hasMore]
      properties:
        changes: { type: array, items: { $ref: '#/components/schemas/SyncChange' } }
        cursor: { type: string, pattern: '^[0-9]+$' }
        hasMore: { type: boolean }
...
    ExternalEntityType:
      type: string
      enum: [track, artist, release]
    ExternalEntityReference:
      type: object
      required: [provider, entityType, externalID]
      properties:
        provider: { type: string }
        entityType: { $ref: '#/components/schemas/ExternalEntityType' }
        externalID: { type: string }
        canonicalURL: { type: [string, 'null'], format: uri }
    ExternalSource:
      type: object
      required: [provider, reference]
      properties:
        provider: { type: string }
        reference: { $ref: '#/components/schemas/ExternalEntityReference' }
    ExternalAudioSource:
      allOf:
        - { $ref: '#/components/schemas/ExternalSource' }
        - type: object
          required: [resolverPath]
          properties: { resolverPath: { type: [string, 'null'] } }
    ExternalAcquisition:
      allOf:
        - { $ref: '#/components/schemas/ExternalSource' }
        - type: object
          required: [method, allowed]
          properties:
            method: { type: string, enum: [direct, yt_dlp, spotdl, unavailable] }
            allowed: { type: boolean }
    MusicSearchItem:
      type: object
      required: [id, entityType, title, duration, metadataSource, acquisition, attribution]
      properties:
        id: { type: string }
        entityType: { $ref: '#/components/schemas/ExternalEntityType' }
        title: { type: string }
        artist: { type: [string, 'null'] }
        release: { type: [string, 'null'] }
        duration: { type: number, minimum: 0 }
        artworkURL: { type: [string, 'null'] }
        metadataSource: { $ref: '#/components/schemas/ExternalSource' }
        audioSource: { anyOf: [{ $ref: '#/components/schemas/ExternalAudioSource' }, { type: 'null' }] }
        acquisition: { $ref: '#/components/schemas/ExternalAcquisition' }
        attribution: { type: string }
    MusicSourceFailure:
      type: object
      required: [provider, kind, message]
      properties:
        provider: { type: string }
        kind: { type: string, enum: [authentication, rate_limit, geo_restricted, temporary, unavailable, malformed_response] }
        message: { type: string }
        retryAfterSeconds: { type: integer, minimum: 0 }
    MusicSearchResponse:
      type: object
      required: [results, failures]
      properties:
        results: { type: array, items: { $ref: '#/components/schemas/MusicSearchItem' } }
        failures: { type: array, items: { $ref: '#/components/schemas/MusicSourceFailure' } }
    AcquisitionRequest:
      type: object
      required: [reference, method, title]
      properties:
        reference: { $ref: '#/components/schemas/ExternalEntityReference' }
        method: { type: string, enum: [direct, yt_dlp, spotdl] }
        title: { type: string }
        artist: { type: [string, 'null'] }
        artworkURL: { type: [string, 'null'], format: uri }
    AcquisitionJob:
      allOf:
        - { $ref: '#/components/schemas/AcquisitionRequest' }
        - type: object
          required: [id, state, progress, attemptCount, maxAttempts, createdAt, updatedAt]
          properties:
            id: { type: string, format: uuid }
            state: { type: string, enum: [queued, running, postProcessing, completed, failed] }
            progress: { type: number, minimum: 0, maximum: 1 }
            attemptCount: { type: integer }
            maxAttempts: { type: integer }
            errorCode: { type: [string, 'null'] }
            errorDetail: { type: [string, 'null'] }
            catalogFileID: { type: [string, 'null'], format: uuid }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    Error:
      type: object
      required: [code, messageKey, requestID]
      properties:
        code: { type: string }
        messageKey: { type: string }
        requestID: { type: string }
        details: {}